Update rails2_ruby2 dependency to point to the chores/postgresql_patches branch
and to update how rails2_ruby2 patches are loaded on a per-component basis.

The change in the require(s) in this commit are to allow each component to be responsible for loading their specific patches. This helps. avoid loading issues where active_support loads the rails2_ruby2 before other components are loaded (e.g. ActiveRecord) which results in the ActiveRecord patches never being loaded.